+++ apps/codeimage/src/components/Frame/Frame.css.ts
@@ -15,7 +15,7 @@ export const [frame, frameVars] = createTheme({
   width: '730px',
   minWidth: '640px',
   minHeight: '150px',
-  maxWidth: '920px',
+  maxWidth: '1400px',
   controlHandleSize: '24px',
   controlHandleColor: themeVars.dynamicColors.frameDragControlBackgroundColor,
   controlOffset: '0px',
@@ -25,8 +25,6 @@ export const wrapper = style({
   width: '100%',
   height: '100%',
   display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
   overflowY: 'auto',
   overflowX: 'hidden',
   flex: '1',
@@ -39,13 +37,23 @@ export const handler = style([
     transform: `scale(${frameHandlerVars.scale})`,
     display: 'block',
     position: 'relative',
+    margin: 'auto',
   },
 ]);
+
+export const content = style({
+  width: '100%',
+  height: '100%',
+  marginTop: '20px',
+  marginBottom: '40px',
+  position: 'relative',
+});
 
 export const container = style([
   frame,
   {
     width: frameVars.width,
+    maxWidth: frameVars.maxWidth,
     minWidth: frameVars.minWidth,
     minHeight: frameVars.minHeight,
     position: 'relative',